Re-enables disabled vex code
If you don't want to use some code, just don't use it. Commenting it out and then re-enabling it all the time clogs up our repo. Having the method there and not using it is just fine. Enable and disable the device in the init method, don't just remove all references of it.