Hey guys, this is Justin. Fixed: multipliers[] in setMovement() was always normalised, so that's why the robot was so hard to do slight turns.
Added:
scale argument in setMovement() if you guys needed the speed to be further decreased for turning.
Driver can toggle between faster speed and scaled down speed using A button (obviously can be changed).
Colour sensor readings in telemetry.  <- this needs to be tested and get the values from it to use in autonomous.
AutonomousMode enum to later replace the if-elseif....else statement in VelAutonomousProgram.
AutonomousStates enum if you guys want to do a step-by-step approach to decrease repeated code.
Added, but commented out background change on the driver station when shooter RPM is near target RPM.  <- not sure it works, but you can give it try.