Reverses direction of goTicks
Reversed the direction that the robot assumes it's going.
When the motor is turning normally, "forward" power makes the shaft turn clockwise when looking down the shaft. Our robot is set up so that "forward" should turn the motors counter-clockwise. So, I reversed the comparison direction on the statements that decide when we stop moving.